Improves user authorization functionalities

+++ app/Http/Controllers/Authorization/RoleController.php
@@ -6,17 +6,22 @@ use App\Models\Role;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\RoleResource;
+use Illuminate\Support\Facades\Gate;
 use Symfony\Component\HttpFoundation\Response;
 
 class RoleController extends Controller
 {
     public function index(){
+        Gate::authorize('view', 'roles');
+
         return response()->json([
-            'roles' => RoleResource::collection(Role::paginate())
+            'roles' => RoleResource::collection(Role::with('permissions')->paginate())
         ], Response::HTTP_OK);
     }
 
     public function show($id){
+        Gate::authorize('view', 'roles');
+
         $role = Role::findOrFail($id);
         
         return response()->json([
@@ -25,34 +30,54 @@ class RoleController extends Controller
     }
 
     public function store(Request $request){
+        Gate::authorize('edit', 'roles');
+
         $request ->validate([
             'name' => ['required', 'unique:roles,name']
         ]);
 
         $role = Role::create($request->all());
 
+        if($permissions = $request->permissions){
+            foreach($permissions as $permission){
+                $role->permissions()->attach($permission);
+            }
+        }    
+
         return response()->json([
             'message' => 'Role '.$role->name.' has been creted successfully',
-            'role' => new RoleResource($role)
+            'role' => new RoleResource(Role::with('permissions')->findOrFail($role->id))
         ], Response::HTTP_CREATED);
     }
 
     public function update($id, Request $request){
+        Gate::authorize('edit', 'roles');
+
         $request->validate([
-            'name' => ['required', 'unique:roles,name,'.$id]
+            'name' => ['unique:roles,name,'.$id]
         ]);
 
         $role = Role::findOrFail($id);
         $role->update($request->only('name'));
 
+        if($permissions = $request->permissions)
+            $role->permissions()->sync($permissions);
+        else
+            $role->permissions()->sync([]);
+        
+
         return response()->json([
             'message' => 'Role has been updated successfully',
-            'role' => new RoleResource($role)
+            'role' => new RoleResource(Role::with('permissions')->findOrFail($role->id))
         ], Response::HTTP_ACCEPTED);
     }
 
     public function destroy($id){
+        Gate::authorize('edit', 'roles');
+
         $role = Role::findOrFail($id);
+        $role->permissions()->detach($role->permissions);
+        
         $role_name = $role->name;
         $role->delete();
